# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thingdex API",
    "description": "Household inventory API. See `/docs` for interactive examples and `/openapi.json` for machine-readable specs.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/label-profiles": {
      "get": {
        "tags": [
          "label-profiles"
        ],
        "summary": "List Label Profiles",
        "operationId": "list_label_profiles_v1_label_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabelProfileOut"
                  },
                  "type": "array",
                  "title": "Response List Label Profiles V1 Label Profiles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "label-profiles"
        ],
        "summary": "Create Label Profile",
        "operationId": "create_label_profile_v1_label_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/label-profiles/{profile_id}": {
      "get": {
        "tags": [
          "label-profiles"
        ],
        "summary": "Get Label Profile",
        "operationId": "get_label_profile_v1_label_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "label-profiles"
        ],
        "summary": "Update Label Profile",
        "operationId": "update_label_profile_v1_label_profiles__profile_id__patch",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "label-profiles"
        ],
        "summary": "Delete Label Profile",
        "operationId": "delete_label_profile_v1_label_profiles__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelProfileCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entity_kind": {
            "type": "string",
            "enum": [
              "item",
              "location"
            ],
            "title": "Entity Kind"
          },
          "item_type_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Type Id"
          },
          "location_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Kind"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "printer_id": {
            "type": "string",
            "title": "Printer Id"
          },
          "auto_print": {
            "type": "boolean",
            "title": "Auto Print",
            "default": true
          },
          "bindings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Bindings"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "entity_kind",
          "template_id",
          "printer_id"
        ],
        "title": "LabelProfileCreate"
      },
      "LabelProfileOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entity_kind": {
            "type": "string",
            "enum": [
              "item",
              "location"
            ],
            "title": "Entity Kind"
          },
          "item_type_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Type Id"
          },
          "location_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Kind"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "printer_id": {
            "type": "string",
            "title": "Printer Id"
          },
          "auto_print": {
            "type": "boolean",
            "title": "Auto Print",
            "default": true
          },
          "bindings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Bindings"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "entity_kind",
          "template_id",
          "printer_id",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "LabelProfileOut"
      },
      "LabelProfileUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "printer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printer Id"
          },
          "auto_print": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Print"
          },
          "bindings": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bindings"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "LabelProfileUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}